add check to block user delete others like

+++ controllers/likeController.js
@@ -1,11 +1,15 @@
 const { Like } = require('../models')
-// const helpers = require('../_helpers')
+const helpers = require('../_helpers')
 
 const likeController = {
   postUnlike: async (req, res) => {
     try {
-      // should I check if this owned by the user?
-      await Like.destroy({ where: { id: Number(req.params.id) } })
+      const targetLike = await Like.findByPk(Number(req.params.id))
+      if (targetLike.toJSON().UserId === helpers.getUser(req).id) {
+        await Like.destroy({ where: { id: Number(req.params.id) } })
+      } else {
+        return res.status(401).json({ status: 'error', message: '無法變更其他使用者的Profile' })
+      }
       return res.status(200).json({ status: 'success', message: '成功取消喜歡' })
     } catch (error) {
       console.log(error)
